Adiciona conexão da página de login do admin com a API, altera localização dos componentes: footer, navbar e SeatPicker e altera localização e posição das páginas de login e do dashboard do admin
