+++ package.json
@@ -17,8 +17,10 @@
     "lucide-react": "^0.554.0",
     "mongodb": "^7.0.0",
     "next": "16.0.3",
+    "next-themes": "^0.4.6",
     "react": "19.2.0",
     "react-dom": "19.2.0",
+    "sonner": "^2.0.7",
     "tailwind-merge": "^3.4.0"
   },
   "devDependencies": {
